# Optimize Base62.compatible?