Commit cac33be introduced a great benchmark to compare different
implementations for `Base62.compatible?`.

I noticed two things:

1. We ran `Enumerable.include?` over a slightly long String, so I thought
converting it to a Set would help - and it does! It's over 3x faster
than the previously fastest method. Please see the benchmark results below.
2. It seems that the chosen implementation differs from the actual
benchmark. I added the code used in `Base62.compatible?` to the
benchmark and it actually seems the slowest of all.

```
Warming up --------------------------------------
            include?    18.811k i/100ms
  include? optimized   113.581k i/100ms
       Regexp#match?    35.244k i/100ms
Regexp#match? - commit cac33be
                         2.932k i/100ms
             Array#-    14.629k i/100ms
          two finger     5.644k i/100ms
Calculating -------------------------------------
            include?    183.872k (± 2.8%) i/s -    921.739k in   5.016951s
  include? optimized      1.088M (± 6.1%) i/s -      5.452M in   5.035707s
       Regexp#match?    339.933k (± 6.7%) i/s -      1.692M in   5.004362s
Regexp#match? - commit cac33be
                         29.828k (± 2.1%) i/s -    149.532k in   5.015367s
             Array#-    139.578k (± 8.6%) i/s -    702.192k in   5.086914s
          two finger     55.528k (± 2.4%) i/s -    282.200k in   5.085177s

Comparison:
  include? optimized:  1087870.6 i/s
       Regexp#match?:   339933.4 i/s - 3.20x  (± 0.00) slower
            include?:   183871.7 i/s - 5.92x  (± 0.00) slower
             Array#-:   139578.3 i/s - 7.79x  (± 0.00) slower
          two finger:    55527.9 i/s - 19.59x  (± 0.00) slower
          Regexp#match? - commit cac33be:    29828.4 i/s - 36.47x  (± 0.00) slower
```