@michaelherold 2022-07-23
Thanks for doing all this legwork! I swear I tested a Set when I made the original change but perhaps I flubbed something.

I plan to get to this this week. I'd like to check it against various Rubies and double check the implementation. Thanks again!